+++ libs/java/dal/db/dal-db/src/main/java/ca/bc/gov/tno/dal/db/entities/DataLocation.java
@@ -99,7 +99,7 @@ public class DataLocation extends AuditColumns {
    * @param name    Unique name
    * @param version Row version value
    */
-  public DataLocation(int id, String name, int version) {
+  public DataLocation(int id, String name, long version) {
     this(id, name);
     this.setVersion(version);
   }
