+++ libs/java/dal/db/dal-db/src/main/java/ca/bc/gov/tno/dal/db/entities/MediaType.java
@@ -100,7 +100,7 @@ public class MediaType extends AuditColumns {
    * @param name    Unique name
    * @param version Row version value
    */
-  public MediaType(int id, String name, int version) {
+  public MediaType(int id, String name, long version) {
     this(id, name);
     this.setVersion(version);
   }
