+++ libs/java/dal/db/dal-db/src/main/java/ca/bc/gov/tno/dal/db/entities/Series.java
@@ -105,7 +105,7 @@ public class Series extends AuditColumns {
    * @param name    Unique name
    * @param version Row version value
    */
-  public Series(int id, String name, int version) {
+  public Series(int id, String name, long version) {
     this(id, name);
     this.setVersion(version);
   }
